Added tape-like loader option and improve autorun handling
The tape-like loader loads everything in the P file as-is and so doesn't
need to set system variables, unless it is overriding the autorun setting.
This loader is much shorter than the original, so I can go back and make
the rom vars moveable to take up the slack and move the dataOffset up a
good amount to at least make up for the system variables.

This loader allows the VU-FILE and VU-CALC programs to work as they had a
checksum routine that included the system variables. The HotZ II assemblers
are large, and so require the second ROM at 32-40K, but most of them may
use that area, so only the 16K RAM version seems to work. The new loader
helps these as some needed info is in the display file and some store the
autostart BASIC line in the system vars!