First try at compressing the loaders
Make the loader vars only what is needed and move to right after the
loader code. Re-arrange so common ones are first. Remove some zero
length block checks to save a few bytes as the ldir will just copy zero
bytes. The dataOffset and sizeLimit will be variable now. The loader var
names are relative to loaderSize now. Also, set NXTLIN and CH_ADD in the
system vars area when -a is used with the tape loader. The shorter tape
loader allows Impact.p to fit in 16K and it works without forcing
autorun to line 4. However, it's now broken vu-file.