Put back zero block length checks, other small fixes
I forgot that if BC is zero that LDIR treats it as 65536 rather than zero.
So, I can't get rid of the zero check, and I added one to the first
program block for when the program in empty and there are only variables.
Moved the block for autoaddr==dfile down as it was blocking the -a
override on a non-autorunning P file. This version is doing pretty well,
except the tape loader is not working on EightyOne's TS1500 mode.